refactor: separate EditClassroomHeaderbar component
